Login security
decided that usernames and emails should be unique and passwords can be repeated. This is because when creating an account on the site the user will be given feedback about what they messed up and if I where to tell them anything that could illude  the fact that the password they used exists it would harm the security of accounts sense users have public usernames and emails which could be attempted in conjunction with the password using bruit force methods.