admin view and profiles functions
Made it so when a session id is linked to the admin table the user sees an employees tab in the header. This will soon implement the advanced employees from class. Also added user profiles tab that appears for every user that is logged in and functions that will allow them to change their preceded activity, delete, and get info about their account.